small refactoring #2
