+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/SaveApplication.kt
@@ -12,6 +12,7 @@ import java.nio.ByteBuffer
 typealias ByteArrayResponse = ResponseEntity<ByteArray>
 typealias StringResponse = ResponseEntity<String>
 typealias EmptyResponse = ResponseEntity<Void>
+typealias KotlinEmptyResponse = ResponseEntity<Unit>
 typealias ByteBufferFluxResponse = ResponseEntity<Flux<ByteBuffer>>
 typealias ResourceResponse = ResponseEntity<Resource>
